TINY-8853: An experiment with two sinks
There are two parts of this spike

* using two sinks
* changing docking to work with absolute positions in scrollable containers, rather than just fixed

Docking changed quite considerably. It's possible that it would be better to just calculate the right positions
for fixed, and use that instead. Both of them require constant updating anyway.

There are many things that were not done. These include

* the sinks passed to inline dialogs should be popup sinks, not dialog sinks
* the sinks passed to tooltips should probably be dependent on what/where the element is
* the configuration needs to be consistently used

Mainly, this showed that it isn't a massive amount of work to get the two sinks to work (and therefore, have
no disconnect between the menu and editor when scrolling inside a scrollable container, but getting docking
to work nicely requires a lot of trade-offs.